requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
PyGithub
